+++ app/Controllers/Auth/LoginController.php
@@ -3,6 +3,7 @@
 namespace App\Controllers\Auth;
 
 use App\Controllers\Controller;
+use App\Web\ValidationChecker;
 use Psr\Http\Message\ResponseInterface as Response;
 use Psr\Http\Message\ServerRequestInterface as Request;
 
@@ -25,6 +26,7 @@ class LoginController extends Controller
 
         return view()->render($response, 'auth/login.twig', [
             'register_enabled' => $this->getSetting('register_enabled', 'off'),
+            'recaptcha_site_key' => $this->getSetting('recaptcha_enabled') === 'on' ? $this->getSetting('recaptcha_site_key') : null,
         ]);
     }
 
@@ -38,21 +40,34 @@ class LoginController extends Controller
      */
     public function login(Request $request, Response $response): Response
     {
+        if ($this->getSetting('recaptcha_enabled') === 'on') {
+            $recaptcha = json_decode(file_get_contents('https://www.google.com/recaptcha/api/siteverify?secret='.$this->getSetting('recaptcha_secret_key').'&response='.param($request, 'recaptcha_token')));
+
+            if ($recaptcha->success && $recaptcha->score < 0.5) {
+                $this->session->alert(lang('recaptcha_failed'), 'danger');
+                return redirect($response, route('login'));
+            }
+        }
+
         $username = param($request, 'username');
         $user = $this->database->query('SELECT `id`, `email`, `username`, `password`,`is_admin`, `active`, `current_disk_quota`, `max_disk_quota` FROM `users` WHERE `username` = ? OR `email` = ? LIMIT 1', [$username, $username])->fetch();
 
-        if (!$user || !password_verify(param($request, 'password'), $user->password)) {
-            $this->session->alert(lang('bad_login'), 'danger');
-            return redirect($response, route('login'));
-        }
+        $validator = ValidationChecker::make()
+            ->rules([
+                'login' => $user && password_verify(param($request, 'password'), $user->password),
+                'maintenance' => !isset($this->config['maintenance']) || !$this->config['maintenance'] || $user->is_admin,
+                'user_active' => $user->active,
+            ])
+            ->onFail(function ($rule) {
+                $alerts = [
+                    'login' => lang('bad_login'),
+                    'maintenance' => lang('maintenance_in_progress'),
+                    'user_active' => lang('account_disabled'),
+                ];
 
-        if (isset($this->config['maintenance']) && $this->config['maintenance'] && !$user->is_admin) {
-            $this->session->alert(lang('maintenance_in_progress'), 'info');
-            return redirect($response, route('login'));
-        }
-
-        if (!$user->active) {
-            $this->session->alert(lang('account_disabled'), 'danger');
+                $this->session->alert($alerts[$rule], $rule === 'maintenance' ? 'info' : 'danger');
+            });
+        if ($validator->fails()) {
             return redirect($response, route('login'));
         }